+++ src/main/java/pl/asie/computronics/block/BlockTapeReader.java
@@ -18,7 +18,7 @@ public class BlockTapeReader extends BlockMachineSidedIcon {
 	private IIcon mFront;
 
 	public BlockTapeReader() {
-		super();
+		super("tape_drive");
 		this.setBlockName("computronics.tapeDrive");
 		this.setGuiProvider(Computronics.guiTapeDrive);
 	}