+++ src/main/java/pl/asie/computronics/integration/buildcraft/pluggable/ItemDroneStation.java
@@ -12,11 +12,12 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
 import pl.asie.computronics.Computronics;
+import pl.asie.computronics.oc.manual.IItemWithPrefix;
 
 /**
  * @author Vexatos
  */
-public class ItemDroneStation extends Item implements IPipePluggableItem {
+public class ItemDroneStation extends Item implements IPipePluggableItem, IItemWithPrefix {
 
 	public ItemDroneStation() {
 		super();
@@ -43,11 +44,21 @@ public class ItemDroneStation extends Item implements IPipePluggableItem {
 
 	@Override
 	public PipePluggable createPipePluggable(IPipe pipe, ForgeDirection side, ItemStack stack) {
-		switch(side){
+		switch(side) {
 			case UP:
 				return new DroneStationPluggable();
 			default:
 				return null;
 		}
+	}
+
+	@Override
+	public String getDocumentationName(ItemStack stack) {
+		return "drone_station";
+	}
+
+	@Override
+	public String getPrefix(ItemStack stack) {
+		return "buildcraft/";
 	}
 }