+++ src/main/java/pl/asie/computronics/integration/railcraft/block/BlockDigitalReceiverBox.java
@@ -10,6 +10,7 @@ import mods.railcraft.common.blocks.signals.BlockSignalBase;
 import mods.railcraft.common.blocks.signals.ISignalTileDefinition;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.texture.IIconRegister;
+import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
@@ -18,6 +19,7 @@ import net.minecraftforge.common.util.ForgeDirection;
 import pl.asie.computronics.Computronics;
 import pl.asie.computronics.integration.railcraft.SignalTypes;
 import pl.asie.computronics.oc.block.IComputronicsEnvironmentBlock;
+import pl.asie.computronics.oc.manual.IBlockWithPrefix;
 import pl.asie.computronics.reference.Mods;
 import pl.asie.computronics.integration.railcraft.tile.TileDigitalReceiverBox;
 
@@ -27,7 +29,7 @@ import pl.asie.computronics.integration.railcraft.tile.TileDigitalReceiverBox;
 @Optional.InterfaceList({
 	@Optional.Interface(iface = "pl.asie.computronics.oc.block.IComputronicsEnvironmentBlock", modid = Mods.OpenComputers)
 })
-public class BlockDigitalReceiverBox extends BlockSignalBase implements IComputronicsEnvironmentBlock {
+public class BlockDigitalReceiverBox extends BlockSignalBase implements IComputronicsEnvironmentBlock, IBlockWithPrefix {
 
 	public static IIcon[] texturesBox;
 	public static IIcon texturesBoxTop;
@@ -111,5 +113,28 @@ public class BlockDigitalReceiverBox extends BlockSignalBase implements IComputr
 	@Optional.Method(modid = Mods.OpenComputers)
 	public Class<? extends Environment> getTileEntityClass(int meta) {
 		return TileDigitalReceiverBox.class;
+	}
+
+	private final String documentationName = "digital_receiver_box";
+	private final String prefix = "railcraft/";
+
+	@Override
+	public String getDocumentationName(World world, int x, int y, int z) {
+		return this.documentationName;
+	}
+
+	@Override
+	public String getDocumentationName(ItemStack stack) {
+		return this.documentationName;
+	}
+
+	@Override
+	public String getPrefix(World world, int x, int y, int z) {
+		return this.prefix;
+	}
+
+	@Override
+	public String getPrefix(ItemStack stack) {
+		return this.prefix;
 	}
 }