+++ src/main/java/pl/asie/computronics/integration/railcraft/block/BlockLocomotiveRelay.java
@@ -6,22 +6,24 @@ import cpw.mods.fml.relauncher.SideOnly;
 import li.cil.oc.api.network.Environment;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.ChatComponentTranslation;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.World;
 import pl.asie.computronics.block.BlockPeripheral;
-import pl.asie.computronics.reference.Mods;
 import pl.asie.computronics.integration.railcraft.tile.TileLocomotiveRelay;
+import pl.asie.computronics.oc.manual.IBlockWithPrefix;
+import pl.asie.computronics.reference.Mods;
 
 /**
  * @author Vexatos
  */
-public class BlockLocomotiveRelay extends BlockPeripheral {
+public class BlockLocomotiveRelay extends BlockPeripheral implements IBlockWithPrefix {
 	private IIcon mTop, mSide, mBottom;
 
 	public BlockLocomotiveRelay() {
-		super();
+		super("locomotive_relay");
 		this.setIconName("computronics:machine_top");
 		this.setBlockName("computronics.locomotiveRelay");
 		this.setNoNedoComputers(true);
@@ -77,5 +79,17 @@ public class BlockLocomotiveRelay extends BlockPeripheral {
 		}
 
 		return super.onBlockActivated(world, x, y, z, player, a, _x, _y, _z);
+	}
+
+	private final String prefix = "railcraft/";
+
+	@Override
+	public String getPrefix(World world, int x, int y, int z) {
+		return this.prefix;
+	}
+
+	@Override
+	public String getPrefix(ItemStack stack) {
+		return this.prefix;
 	}
 }