+++ src/main/java/pl/asie/computronics/oc/manual/IItemWithDocumentation.java
@@ -1,0 +1,11 @@
+package pl.asie.computronics.oc.manual;
+
+import net.minecraft.item.ItemStack;
+
+/**
+ * @author Vexatos
+ */
+public interface IItemWithDocumentation {
+
+	public String getDocumentationName(ItemStack stack);
+}