+++ src/main/java/pl/asie/computronics/tile/TileEEPROMReader.java
@@ -84,7 +84,7 @@ public class TileEEPROMReader extends TileEntityPeripheralBase implements IInven
 		} else return null;
     }
 
-    @Callback(doc = "function(pos:number, data:number or string):boolean;Writes the specified data to the specified position. Returns true on success", direct = true)
+    @Callback(doc = "function(pos:number, data:number or string):boolean; Writes the specified data to the specified position. Returns true on success", direct = true)
     @Optional.Method(modid=Mods.OpenComputers)
     public Object[] write(Context context, Arguments args) {
     	byte[] data = getData();